list-view: Reimplement using GtkColumnView
GtkTreeView, while still available in GTK 4, is more limited in some
more specialized situations which we have been relying on, such as
drag-and-drop and high DPI icons.

It's also been our long desire to adopt GListModel-based list widgets
in order to unlock features and bugfixes which would have been
basically impossible to do with GtkTreeView.

We are thus dropping GtkTreeView for good and adopting GtkColumnView.

The new implementation is radically different; almost no code remains
from the old implementation. However, the new implementation has full
feature parity with the old one with two exceptions:

    1. Expand subfolders as a tree: WIP in another branch.
    2. Performance for large number of items: WIP branch in GTK.

Same as the old implementation, it still lacks drag-and-drop support,
which is yet to be reimplemented for GTK4.

The new implementation also implements new features which were but
a dream in GtkTreeView:

  - Rubberband-selection.
  - Empty space inbetween and around the list of items to open context
    menu, start rubberband selection, drop items, clear selection...
  - File names in search results and recents are no longer squashed by
    the "Original location" column containing long paths. Instead, the
    original location is runs parallel to the filename.
  - With the location column gone, the size column can be displayed
    again in these two special locations.
  - Full-text-search snippets no longer compete for horizontal space
    with the filenames, but are displayed as accent-colored subtitles.
  - Sort order can be changed from the view menu, same as in grid view.
  - Per-folder sorting is shared with the grid view now, fixing an old
    inconsistency.
  - A starring animation ☆★

Closes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/320