list-view: Hide expander if subdirectory is empty
The old list view used to add a dummy row with the "(empty)" label.

That solution required a lot of code gymnastics to special case
the dummy row everywhere, which was error-prone.

So, for the new list view, I wanted to avoid that solution. For
now, the simplest solution is to follow the example of other
file managers and hide the expander if the folder is found to be
empty after trying to expand it.