view-icon-controller: Abstract sharable code
The new list view is going to be GtkColumnView-based, so it's going to
share some code with the GtkGridView-based view.

In order to avoid code duplication and still keep the NautilusFilesView
class agnostic of the widgets used by final classes, create an abstract
NautilusListBase class.

But this abstract class needs to interact with the item widgets, which
are going to be different between views. To resolve this, an abstract
NautilusViewCell class is created for the item widgets, which is also
going to be used for the new list view column cells.

Also, bump GLib version requirement now that we use GSignalGroup.